requirements.txt
appdirs==1.4.3
httplib2==0.10.3
httpretty==0.8.14
imageio==2.1.2
numpy==1.12.0
oauth2==1.9.0.post1
oauthlib==2.0.1
olefile==0.44
packaging==16.8
pbr==2.0.0
Pillow==4.0.0
pyparsing==2.2.0
PyTumblr==0.0.6
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
stevedore==1.21.0
tweepy==3.5.0
